@@ -15,13 +15,13 @@
     "cdk": "cdk"
   },
   "devDependencies": {
-    "@types/node": "^8.9.4",
-    "typescript": "^3.1.2",
-    "aws-cdk": "^0.24.1"
+    "@types/node": "^12.6.1",
+    "typescript": "^3.5.3",
+    "aws-cdk": "^0.39.0"
   },
   "dependencies": {
-    "@aws-cdk/aws-ec2": "^0.24.1",
-    "@aws-cdk/cdk": "^0.24.1",
-    "source-map-support": "^0.5.9"
+    "@aws-cdk/aws-ec2": "^0.39.0",
+    "@aws-cdk/core": "^0.39.0",
+    "source-map-support": "^0.5.12"
   }
 }
